next-app(refactor): colocate components in routes under (components)/
