+++ packages/next-app/src/app/(components)/RootLayout.tsx
@@ -6,13 +6,13 @@ import { useRouter } from "next/navigation";
 import { useContext, useEffect } from "react";
 import { useAccount, useNetwork, WagmiConfig } from "wagmi";
 
-import { OwnedPool } from "../components/OwnedPool";
 import {
   PoolMetadataContext,
   PoolMetadataProvider,
-} from "../contexts/PoolMetadataContext";
-import gql from "../lib/gql";
-import { chains, client } from "../wagmi";
+} from "../../contexts/PoolMetadataContext";
+import gql from "../../lib/gql";
+import { chains, client } from "../../wagmi";
+import { OwnedPool } from "./OwnedPool";
 
 export function RootLayout({ children }: { children: React.ReactNode }) {
   const { chain } = useNetwork();
