+++ packages/next-app/src/app/layout.tsx
@@ -4,7 +4,7 @@ import "../styles/global.css";
 import type { Metadata } from "next";
 import * as React from "react";
 
-import { RootLayout } from "./RootLayout";
+import { RootLayout } from "./(components)/RootLayout";
 
 export const metadata: Metadata = {
   title: "Balancer Pool Metadata",
