+++ packages/next-app/src/app/pool/[poolId]/page.tsx
@@ -1,9 +1,9 @@
-import { MetadataAttribute } from "./MetadataAttribute";
+import { MetadataAttributes } from "./(components)/MetadataAttributes";
 
 export default function Page({ params }: { params: { poolId: string } }) {
   return (
     <div className="h-full flex-1 py-5 text-white">
-      <MetadataAttribute poolId={params.poolId} />
+      <MetadataAttributes poolId={params.poolId} />
     </div>
   );
 }
