+++ packages/next-app/src/components/index.ts
@@ -1,4 +1,2 @@
 export { Button } from "./Button";
 export { ImageDialog } from "./ImageDialog";
-export { OwnedPool } from "./OwnedPool";
-export { PoolMetadataFormModal } from "./PoolMetadataFormModal";
